+++ NEWS.md
@@ -8,6 +8,8 @@
 
 - Now the user is able to restrict the editing to accept only numbers by e.g. `datatable(..., editable = list(target = "cell", numeric = 3:5))`. By default the editing is restricted to numbers for all numeric columns; set `numeric = "none"` to disable this behavior (#824).
 
+- Now the user can request text areas for the editing by e.g. `datatable(..., editable = list(target = "cell", area = 1:2))`. Text areas are useful for cells with large contents.
+
 
 ## MINOR CHANGES